Fix nibabel dtype compatibility for int64 and float arrays

+++ vox2organ/utils/evaluate.py
@@ -115,7 +115,7 @@ class ModelEvaluator:
             if run_model:
                 # Generate prediction and store
                 input_img = nib.Nifti1Image(
-                    data["img"].cpu().numpy(),
+                    data["img"].cpu().numpy().astype(np.float32),
                     np.eye(4)
                 )
                 if logger.debug():
@@ -127,7 +127,7 @@ class ModelEvaluator:
                         )
                     )
                 voxel_label = nib.Nifti1Image(
-                    data["voxel_label"].cpu().numpy(),
+                    data["voxel_label"].cpu().numpy().astype(np.int32),
                     np.eye(4)
                 )
                 if logger.debug():
